On-device onroad CI test (#19792)
* start onroad test

* cpu usage test works

* nice print

* run in jenkins

* pretty reliable

* bump loggerd cpu usgae
