+++ release/files_common
@@ -333,7 +333,6 @@ selfdrive/thermald/power_monitoring.py
 selfdrive/test/__init__.py
 selfdrive/test/helpers.py
 selfdrive/test/setup_device_ci.sh
-selfdrive/test/test_cpu_usage.py
 selfdrive/test/test_fingerprints.py
 selfdrive/test/test_manager.py
